refactor: update component import paths across all dashboard pages
Adjust DashboardLayout imports across all dashboard route pages to use the new @/components/layout/ path, also update BrandList import location. This fixes broken module resolution after project component directory restructuring.
