+++ src/app/dashboard/admin/settings/page.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect } from "react";
 import { useAuth } from "@/lib/auth";
 import { useRouter } from "next/navigation";
-import DashboardLayout from "@/components/DashboardLayout";
+import DashboardLayout from "@/components/layout/DashboardLayout";
 
 // Admin role ID
 const ADMIN_ROLE_ID = 1;
