+++ src/app/dashboard/brands/create/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import { useRouter } from "next/navigation";
-import DashboardLayout from "@/components/DashboardLayout";
+import DashboardLayout from "@/components/layout/DashboardLayout";
 import { CreateBrandSchema } from "@/lib/validations";
 import type { z } from "zod";
