+++ src/app/dashboard/datatable-demo/page.tsx
@@ -12,7 +12,7 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
-import DashboardLayout from "@/components/DashboardLayout";
+import DashboardLayout from "@/components/layout/DashboardLayout";
 import { ThemedDataTable } from "@/components/DataTable";
 
 import "@/app/dashboard/datatable-demo/theme.css";
