+++ src/app/dashboard/settings/page.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import { useAuth } from "@/lib/auth";
-import DashboardLayout from "@/components/DashboardLayout";
+import DashboardLayout from "@/components/layout/DashboardLayout";
 
 export default function SettingsPage() {
   const { user, isLoading } = useAuth();
