+++ src/app/dashboard/teams/[id]/page.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect, use, useRef, useCallback } from "react";
 import { useRouter } from "next/navigation";
 import Link from "next/link";
-import DashboardLayout from "@/components/DashboardLayout";
+import DashboardLayout from "@/components/layout/DashboardLayout";
 
 interface TeamMember {
   id: number;
