+++ src/app/dashboard/users/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import UserList from "@/components/UserList";
-import DashboardLayout from "@/components/DashboardLayout";
+import UserList from "@/components/tables/lists/UserList";
+import DashboardLayout from "@/components/layout/DashboardLayout";
 
 export default function UsersPage() {
   const router = useRouter();
