+++ src/components/form/AsyncSelect.tsx
@@ -105,7 +105,7 @@ export function AsyncSelect({
         >
           <option value="">{placeholder}</option>
           {options.map((option) => (
-            <option key={option.id} value={option.id}>
+            <option key={option.id} value={option.name}>
               {option.name}
             </option>
           ))}
