+++ src/components/tables/lists/BrandList.tsx
@@ -1,20 +1,23 @@
 "use client";
 
+import { useState } from "react";
 import {
   ServerSideDataTable,
   FilterConfig,
-} from "@/components/ServerSideDataTable";
-import { columns, User } from "@/components/UserColumns";
+} from "@/components/tables/ServerSideDataTable";
+import { columns, Brand } from "@/components/tables/columns/BrandColumns";
 import { useServerSideDataTable } from "@/hooks/useServerSideDataTable";
 
-interface UserListProps {
-  /** Optional callback when Add User button is clicked */
+interface BrandListProps {
+  /** Optional callback when Create Brand button is clicked */
   onAddClick?: () => void;
 }
 
-export default function UserList({ onAddClick }: UserListProps) {
+export default function BrandList({ onAddClick }: BrandListProps) {
+  const [industries, setIndustries] = useState<string[]>([]);
+
   const {
-    data: users,
+    data: brands,
     loading,
     error,
     pagination,
@@ -26,44 +29,42 @@ export default function UserList({ onAddClick }: UserListProps) {
     setSearch,
     extraParams,
     setExtraParams,
-  } = useServerSideDataTable<User>({
-    apiEndpoint: "/api/users",
+  } = useServerSideDataTable<Brand>({
+    apiEndpoint: "/api/brands",
     defaultSortBy: "createdAt",
     defaultSortOrder: "desc",
+    onDataFetched: (data) => {
+      setIndustries(data.industries || []);
+    },
   });
 
   // Build filters configuration
   const filters: FilterConfig[] = [
-    {
-      id: "role",
-      label: "Role",
-      type: "select",
-      options: [
-        { value: "admin", label: "Admin" },
-        { value: "hr", label: "HR" },
-        { value: "employee", label: "Employee" },
-      ],
-      value: extraParams.role || "",
-      onChange: (value: string) =>
-        setExtraParams((prev) => ({ ...prev, role: value })),
-    },
     {
       id: "status",
       label: "Status",
       type: "select",
       options: [
         { value: "active", label: "Active" },
-        { value: "on_leave", label: "On Leave" },
-        { value: "terminated", label: "Terminated" },
         { value: "inactive", label: "Inactive" },
+        { value: "archived", label: "Archived" },
       ],
       value: extraParams.status || "",
       onChange: (value: string) =>
         setExtraParams((prev) => ({ ...prev, status: value })),
     },
+    {
+      id: "industry",
+      label: "Industry",
+      type: "select",
+      options: industries.map((ind: string) => ({ value: ind, label: ind })),
+      value: extraParams.industry || "",
+      onChange: (value: string) =>
+        setExtraParams((prev) => ({ ...prev, industry: value })),
+    },
   ];
 
-  // Create the Add User button if onAddClick is provided
+  // Create the Create Brand button if onAddClick is provided
   const searchAction = onAddClick ? (
     <button
       onClick={onAddClick}
@@ -82,19 +83,19 @@ export default function UserList({ onAddClick }: UserListProps) {
           d="M12 4.5v15m7.5-7.5h-15"
         />
       </svg>
-      Add User
+      Create Brand
     </button>
   ) : null;
 
   return (
     <ServerSideDataTable
       columns={columns}
-      data={users}
+      data={brands}
       totalCount={pagination.total}
       isLoading={loading}
       error={error}
       searchKey="name"
-      searchPlaceholder="Search by employee name..."
+      searchPlaceholder="Search by brand name..."
       searchValue={search}
       onSearchChange={setSearch}
       searchAction={searchAction}
