+++ src/components/widgets/ClockInButton.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect } from "react";
 import { useAuth } from "@/lib/auth";
 import { useToast } from "@/lib/toast";
-import Modal from "@/components/Modal";
+import Modal from "@/components/modals/Modal";
 
 interface AttendanceStatus {
   attendance: {
